Fix the hotreload dll on linux and also fix the bash scripts on linux

+++ clean.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
+if [[ -e "buildout/" ]]; then
+echo "Deleting folder buildout"
 rm buildout/*.*
 rmdir buildout
+fi
